Merge branch 'deuna-feature' into 'main'

+++ microservice-pos/src/main/java/com/cumpleanos/pos/persistence/api/deuna/payments/PaymentRequest.java
@@ -3,7 +3,7 @@ package com.cumpleanos.pos.persistence.api.deuna.payments;
 public record PaymentRequest(
         String pointOfSale,
         String qrType,
-        long amount,
+        Double amount,
         String detail,
         String internalTransactionReference,
         String format
